use react-query to fetch the homepage modules
