+++ src/features/home/api.test.ts
@@ -5,6 +5,7 @@ import { homepageEntriesAPIResponse, adaptedHomepageEntries } from 'tests/fixtur
 import { server } from 'tests/server'
 
 import { getHomepageEntries, CONTENTFUL_BASE_URL } from './api'
+
 server.use(
   rest.get(
     `${CONTENTFUL_BASE_URL}/spaces/${env.CONTENTFUL_SPACE_ID}/environments/${env.CONTENTFUL_ENVIRONMENT}/entries?include=2&content_type=homepage&access_token=${env.CONTENTFUL_ACCESS_TOKEN}`,
